Refactor diff manager. Optimise performance
