+++ src/utils/index.ts
@@ -2,4 +2,4 @@ export { sanitizeTitle, sanitizeTitleExcess } from './sanitizeTitle';
 export { frontMatter } from './frontmatter';
 export { br2ln } from './br2ln';
 export { hash } from './hash';
-export { sb } from './stringBuffer';
+export { sb, StringBuffer } from './stringBuffer';
